requirements.txt
google-cloud-bigquery==2.0.0
python-dotenv